Solution for ex4.34. Closed #7
The way to output the list contents can be refined
because the way in this commit evaluates only the first argument of `(cons ..)` in `eval`.
It should be evaluated of the real value in the end of `driver-loop` (or `user-print`) after `eval`
rather than in the middle of `eval`.

In addition, the deep elements' evaluation of the nested `(cons .. (cons ...))` expression is done also,
but it is not used anywhere. It's useless.